requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2.1
django-admin-sortable2==2.1.8
django-tinymce==3.6.1
idna==3.4
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.30.0
sqlparse==0.4.4
urllib3==2.0.2
